requirements.txt
yfinance>=0.2.43
loguru>=0.7.2
pydantic>=2.9.2
python-dotenv>=1.0.1
PyYAML>=6.0.2
APScheduler>=3.10.4
jinja2>=3.1.4
feedparser>=6.0.11
openai>=1.54.4
python-telegram-bot>=21.7
